requirements.txt
appium-python-client==1.3.0
selenium==3.141.0
pytest==6.2.5

